+ populate tool with real data + sampled lightbox + db: stored procedure to retrieve tool data + db: revision details table

+++ SFApplication/ngControllers/itemController.js
@@ -11,7 +11,7 @@
                 e.preventDefault();
                 var url = '/Items/EditItem/' + e.target.dataset.itemId;
 
-                popupwindow(url, 300, 600);
+                popupwindow(url, 300, 600, $scope.refreshList);
             });
 
 
@@ -19,6 +19,27 @@
                 e.preventDefault(e.target.dataset.itemId);
                 $scope.deleteItem(e.target.dataset.itemId);
             });
+        });
+    };
+
+    $scope.refreshList = function () {
+        console.log("triggered");
+        $http({
+            method: "GET",
+            url: "/Items/ListItems",
+        }).then(function (response) {
+            $scope.items = response.data;
+
+            //get counts
+            $scope.synced = ($.grep($scope.items, function (item) {
+                return item.Status == "Synced";
+            })).length;
+            $scope.custom = ($.grep($scope.items, function (item) {
+                return item.Status == "Custom";
+            })).length;
+            $scope.imported = ($.grep($scope.items, function (item) {
+                return item.Status == "Imported";
+            })).length;
         });
     };
 
@@ -103,6 +124,11 @@
     };
 
     $scope.deleteItem = function (id) {
+        var confirm_delete = confirm("Are you sure you want to delete this item?");
+        if (confirm_delete != true) {
+            return;
+        }
+
         $http({
             method: "POST",
             url: "/Items/DeleteItem/",
@@ -157,8 +183,6 @@
                 }
             }
         }
-
-        console.log($scope.errors);
     }
 
 });
